bot/requirements.txt
arrow==1.2.3
beautifulsoup4==4.12.2
binaryornot==0.4.4
# Editable install with no version control (bot-alive==1.0)
-e c:\users\sandro bispo\desktop\material_igor\bot\bot_alive
botcity-framework-base==0.4.2
botcity-framework-web==0.8.3
botcity-maestro-sdk==0.3.4
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cookiecutter==2.3.0
distro==1.8.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
msedge-selenium-tools==3.141.3
numpy==1.25.2
opencv-python==4.8.0.76
packaging==23.1
Pillow==10.0.0
Pygments==2.16.1
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.5.2
selenium==3.141.0
six==1.16.0
soupsieve==2.4.1
text-unidecode==1.3
urllib3==1.26.16
zipp==3.16.2
